V0.1.0 - First working demo with cleaned up code (#2)
### **EDIT :** no need to check memory leaks, will be adressed
separaretly


The PR to start anew on a cleaner basis, finally! 

Main goal of _after_ the PR is to scale the project, add new nembers to
the team, and start working on funnier stuff ✨

Lot of work still left to do, and some pretty ugly code/tape still
running in some places; but should be mostly okay to build on.
Two aspects need to be adressed: a technical one, to make sure the code
is ready to be merged, and a more subjective one, as we should use this
milestone to check that the conception is going in the right direction.

**If you only want to make sure everything is working, simply go
throught the _Technical section_. If you want to know more about how
everything works, read everything. Everyone read the last words tho!**

# Project structure

Headers file are in the include folder, source code in src. CMake files
are recursive to each folder to prepare for the complexity of the
project structure. Executables are built to the bin folder - not
included in the git repository.

## Game, Engine, Graphics, Others

I have tried to separate source files in three big folder. The goal is
to split as much as possible each "main system", so we have the option
in the long term to create independant projects and reuse separate
components on other projects.

### Engine

Should contain most of the generic, engine-type stuff. A lot of virtual
classes, the physic engine, and most of the internal logic of the
project.

### Game

Everything that will be game-specific, ie game characters, Entity class
implementations, components implementations, ...
The dev folder give the "spirit" of what should go there.

### Graphics
Not to be a focus for quite some time! Everything fancy that will be
required down the road, including f.e. texture loading, more advanced
displaying, maybe particles, ...

### Others
Currently only keybind loading. But we should probably put there saving,
preferences, options loading as a whole, etc...

# Technical

The following _technical_ things need to be checked before we proceed
with the project:
- **Readme is clear enough**
- **build instructions are correct and clear enough**
- **building works on other platforms** (ie Linux), other compilers...
especially important since ideally, a lot of people are going to work on
the project
- code is clear and clean enough;
- **and** documented enough to be understandable
- **POTENTIAL MEMORY LEAKS**
- _not required_, but look at access control if possible

I believe items in bold are the minimal requirement to accept the PR,
the rest is "somewhat" secondary.

Some precision on code clarity and "cleanliness": everything that has a
**TODO** attached is _not_ meant to be adressed with this PR. We can
still discuss it ofc, but they will be adressed afterwards as issues,
for the following reasons:
- Some stuff require a lot more conception and work, and will stay as-is
for now;
- Some other stuff could be excellent entry-point for new members into
the project, especially for the first years with minimal OOP knowledge.

# Conception

On a more subjective point, we should take some time to discuss whether
or not the project is flexible enough to be build on.
This will require a lot of discussion (that we will keep out of this PR,
unless you see an obvious flaw that can be fixed), but it would be nice
if everyone could think a little bit about this while reviewing the
code.
To help with this, I have attached an elided UML of the current state of
the project at the end of the PR.
I try here to brief on all the subsystems.

## Graphics/Display

Prettry straighforward. Straight pipe between the engine and the SFML
library

## Box

System that will handle intersection of shapes. Currently used for
collisions, soon to be used for the combat system (hitboxes and
hurtboxes f.e.). Only implements rectangles as making several shapes
interacts is a big headache, and is not exactly required as of today.

## Physics

Contains a simple gravity to make objects fall; and a not so simple
subsyste,

## Input

Simply translates keys to entity movements. Very very basic, lacks a lot
of stuff including a jump subsystem, proper buffering of input,
animations, etc...

## Universe Master

I think this is the strongest point of the conception. I believe the
code illustrates by itself the beauty and simplicity of
addind/removing/updating component to the game.

## Entity

Keystone of our Entity-Component-System, yet I fear it is already
outdated. Please pay attention to this class in particular and think
about way to further separate the entities and the components that I
find currently too linked.

## UML diagram

Not complete, lacks a LOT of methods, attributes, access control...


![ClassDiagramAll_22_12_04](https://user-images.githubusercontent.com/12777915/205471348-6b4cb7d8-e1cb-44cd-a282-55e3972c4bc1.jpg)

# Last words

Quite some work done, but more to come! 
Thanks everyone for thrusting me with the project cleanup, and sorry it
took so much time to get there

But I believe the "fil rouge" will finally start for real, 1 year after
we quicklaunched it!

Very lenghty PR, but I think it is coherent with the amount of
modifications. If I forgot something, or you want to do some remark on
the PR (content, form, ...) don't hesitate to do so.

Co-authored-by: Clement <mabileau.clement@gmail.com>
Co-authored-by: Benoit Tellier <benoitprm@pm.e>
Co-authored-by: benoit <tellierbenoit@hotmail.fr>
Co-authored-by: matt <bassetmatt@gmail.com>
Co-authored-by: Xawav <real.xavier2001@gmail.com>
Co-authored-by: alex <alexandre.tullot@gmail.com>
Co-authored-by: Akainoru <PIXELtmGamer@gmail.com>