Improve readme (#21)
Should close #13;

Set CMake requirement to the one already in the main `CMakeLists.txt`;
`sfml` to the current (latest) version; and `cpp` to the not-quite
latest (ie C++17) standard.
Shouldn't have issue by setting sfml to the latest version, but we
should probably enforce `C++17` and CMake requirements in our test suite
and stuff, but that can be addressed by other PR and issues.