Add engine namespace (#28)
Close #24. Basically, add every file under the `src/engine` path to the
namespace `Engine`.

Also add to fix a bunch of other files to fit with the new namespace
(basically just appended `Engine::...` to a bunch of stuff).

If it compiles, it should be good to go and good enough for a review.
Try to check if I missed any file, but should be a-ok.

Choose to go with CamlCase for namespace format, because it looks
prettier imo. Can be changed back to `engine` with a simple Ctrl+F
anyway.